Refactored Piece to an abstract class instead of interface - also added currentLocation to all pieces

+++ src/main/java/com/coolpotato/fun_projects/chess_application/models/Direction.java
@@ -6,11 +6,11 @@ import java.util.Objects;
 
 @Data
 public class Direction {
-    double x;
-    double y;
+    int x;
+    int y;
     private static final double errorMargin = 0.001;
 
-    public Direction(double x, double y) {
+    public Direction(int x, int y) {
         this.x = x;
         this.y = y;
     }
@@ -21,8 +21,8 @@ public class Direction {
         if (o == null || getClass() != o.getClass()) return false;
         Direction other = (Direction) o;
 
-        double xRelation = this.x / other.x;
-        double yRelation = this.y / other.y;
+        double xRelation = (double) this.x / other.x;
+        double yRelation = (double) this.y / other.y;
 
 
         return xRelation / yRelation > 1 - errorMargin && xRelation / yRelation < 1 + errorMargin;
